requirements.txt
pandas
sqlalchemy
influxdb
libimobiledevice